Add conditional formatting for index view
Add conditional formatting in EJS template for index. If the site
visitor isn't a user, usernames and dates will be rendered anonymously
as asterisks. If the visitor is a user and logged in, they will see the
same unless they are members, in which case they can see the full
username and post dates.